Restructure the CLI bootstrap library into concern-scoped modules
commands/bootstrap.ts was a 2,154-line, 36-export shared library with no command of its own, imported by all five real commands. Its concerns now live in dedicated modules: connectionFromUrl.ts (the merged URL-to-connection builder, absorbing zeroSetup's near-identical copy), reconcile.ts (connection diffing), optionDefinitions.ts (shared CLI option definitions and their warnings), util/cli.ts (runOrExit joins the existing CLI helpers), and onlineBootstrap.ts (the online-bootstrap orchestration that remains). commands/ now holds only real command modules. The connection and reconcile modules are deliberately CLI-surface-free, pure over @psilink/core config types, so a future move into the core package is a file move. zeroSetup.ts loses its duplicated builder and its section-banner comments.

One behavior change, from normalizing the single divergence between the two builder copies: a host-less sftp:// URL on the zero-setup path now fails fast with a usage error (exit 64), as it always did on the other commands, instead of reaching the live connection with an empty host; a regression test pins the zero-setup path.

The planned runOrExit unification for exchange and zeroSetup was investigated and deliberately not applied: their scaffolding differences are load-bearing, pinned by tests (bare console.error parse reporting, the missing-config ENOENT-to-64 exit mapping, advisory-before-error ordering), and the CLI machine interface is contract surface.
