+++ apps/cli/src/onlineBootstrap.ts
@@ -1,7 +1,4 @@
-import { fileURLToPath } from "node:url";
-
-import type { Argv, Arguments } from "yargs";
-import logLibrary from "loglevel";
+import type { Arguments } from "yargs";
 
 import {
   getLogger,
@@ -18,10 +15,7 @@ import {
   MAX_ENDPOINT_HOST_LENGTH,
   MAX_ENDPOINT_PATH_LENGTH,
   MAX_PAYLOAD_ENTRIES,
-  normalizeFiledropPath,
   safeParseConnectionConfig,
-  sanitizeErrorForDisplay,
-  MAX_RECONNECT_ATTEMPTS,
   UsageError,
 } from "@psilink/core";
 import type {
@@ -39,56 +33,20 @@ import type {
   WebRTCConnectionConfig,
 } from "@psilink/core";
 
-import {
-  applyConnectionOverrides,
-  saveConfig,
-  type ConnectionOverrides,
-  type ReconcileDiff,
-  RECONCILE_UNSET,
-  DEFAULT_CONFIG_PATH,
-} from "../config";
-import { detectFileConflicts } from "../fileUtils";
-import { DEFAULT_KEY_PATH } from "../keyFile";
-import { resolveConnectionCredentials } from "../util/atSignRefs";
-import {
-  establishHostKeyTrust,
-  type HostKeyPersistence,
-} from "../hostKeyTrust";
-import {
-  durationFlagMs,
-  durationFlagSeconds,
-  LOG_LEVELS,
-  MAX_TIMEOUT_SECONDS,
-  nonNegativeIntFlag,
-  singleValue,
-  openInputSource,
-} from "../util/cli";
-import {
-  DURATION_VALUE_HELP,
-  FINE_DURATION_VALUE_HELP,
-} from "../util/duration";
-import { runProtocol, type AuthPersist } from "../protocol";
-import {
-  decodeUrlComponent,
-  redactUrlCredentials,
-} from "../util/connectionUrl";
-import { channelFromURL } from "./zeroSetup";
-import type { RecordOutput } from "../recordFile";
+import { saveConfig } from "./config";
+import { detectFileConflicts } from "./fileUtils";
+import { resolveConnectionCredentials } from "./util/atSignRefs";
+import { establishHostKeyTrust, type HostKeyPersistence } from "./hostKeyTrust";
+import { openInputSource, singleValue } from "./util/cli";
+import { runProtocol, type AuthPersist } from "./protocol";
+import type { RunnableConnectionConfig } from "./connectionFromUrl";
+import type { RecordOutput } from "./recordFile";
 
 // The exchange-data portion of a spec: linkage terms (always present once
 // resolved) plus the optional metadata and standardization. Distinct from
 // core's ExchangeDataSpec, whose linkageTerms is Partial because it models the
 // not-yet-resolved input to prepareForExchange; here resolution has happened.
 export type ResolvedDataSpec = Omit<ExchangeSpec, "connection">;
-
-// The connection channels the CLI can actually run an exchange over: runProtocol
-// supports sftp and filedrop, and a webrtc URL is rejected upstream. Narrowing
-// to this (rather than the full ConnectionConfig) keeps a webrtc config from
-// reaching runOnlineBootstrap, where it would otherwise only fail at runtime.
-export type RunnableConnectionConfig = Extract<
-  ConnectionConfig,
-  { channel: "sftp" | "filedrop" }
->;
 
 // The placeholder host/username written into a config when the inviter did not
 // supply a connection endpoint. Chosen to be obvious in a diff and to fail
@@ -135,304 +93,6 @@ export function looksLikeUrl(value: string): boolean {
     return false;
   }
   return ["sftp:", "ssh:", "ws:", "wss:", "file:"].includes(url.protocol);
-}
-
-/**
- * Build a connection config from a server URL, for the online invite/accept
- * paths. Mirrors the zero-setup mapping but is constrained to the channels the
- * CLI can actually run: a `webrtc` (ws/wss) URL or an unsupported scheme is a
- * usage error. The returned config carries no `authentication`; the caller adds
- * the shared secret separately for the handshake and never persists it to the
- * config.
- *
- * @internal exported for testing
- */
-export function connectionFromURL(
-  url: URL,
-  overrides: ConnectionOverrides,
-): RunnableConnectionConfig {
-  const channel = channelFromURL(url);
-
-  if (channel === "filedrop") {
-    if (url.hostname && url.hostname !== "localhost")
-      throw new UsageError(
-        `file:// URLs must use three slashes (e.g. file:///mnt/share/drop) ` +
-          `or file://localhost/path; got: ${redactUrlCredentials(url)}`,
-      );
-    const base: FileDropConnectionConfig = {
-      channel: "filedrop",
-      path: fileURLToPath(url),
-    };
-    return applyConnectionOverrides(
-      base,
-      overrides,
-    ) as RunnableConnectionConfig;
-  }
-
-  if (channel !== "sftp")
-    throw new UsageError(`${channel} channel not yet supported in the CLI`);
-
-  // Reject a credential-only or schemeless URL with no host (e.g. sftp:///path)
-  // here, with a clear message, rather than passing host: "" through to a
-  // connection attempt that fails obscurely later. Mirrors the filedrop branch's
-  // host validation above. (redactUrlCredentials is defensive consistency: a
-  // host-less URL cannot actually carry credentials -- the parser rejects
-  // userinfo without a host -- but URLs are always echoed through the redactor.)
-  if (!url.hostname)
-    throw new UsageError(
-      `sftp URL must include a host (e.g. sftp://host/path); got: ` +
-        redactUrlCredentials(url),
-    );
-
-  const base: SFTPConnectionConfig = {
-    channel: "sftp",
-    server: {
-      host: decodeUrlComponent(url.hostname, url),
-      port: url.port ? Number(url.port) : undefined,
-      username: url.username
-        ? decodeUrlComponent(url.username, url)
-        : undefined,
-      password: url.password
-        ? decodeUrlComponent(url.password, url)
-        : undefined,
-      // A bare-host URL (sftp://host or sftp://host/) leaves the remote path
-      // unset so the server's default working directory is used, rather than
-      // pinning it to the filesystem root.
-      path:
-        url.pathname && url.pathname !== "/"
-          ? decodeUrlComponent(url.pathname, url)
-          : undefined,
-    },
-  };
-  return applyConnectionOverrides(base, overrides) as RunnableConnectionConfig;
-}
-
-// The port an SFTP connection uses when the config sets none (ssh2's default).
-// A config with no port and a target stating this value describe the same
-// endpoint, so the reconcile must not flag that as a divergence.
-const DEFAULT_SFTP_PORT = 22;
-
-// Two hosts are the same endpoint regardless of case (DNS is case-insensitive),
-// so compare them case-folded. Paths are compared the way the live connection
-// will treat them, so the reconcile does not abort on a difference the
-// connection would not see -- but the two channels normalize differently, so
-// each has its own comparator. FileSyncConnection.open strips at most one
-// trailing slash from an sftp remote path, while a filedrop path additionally
-// has backslashes folded to forward slashes and ALL trailing slashes stripped.
-function hostsEqual(a: string, b: string): boolean {
-  return a.toLowerCase() === b.toLowerCase();
-}
-function sftpPathsEqual(a: string | undefined, b: string | undefined): boolean {
-  const strip = (p: string | undefined): string | undefined =>
-    p !== undefined && p.endsWith("/") ? p.slice(0, -1) : p;
-  return strip(a) === strip(b);
-}
-function filedropPathsEqual(
-  a: string | undefined,
-  b: string | undefined,
-): boolean {
-  // Normalize both sides through the connection's own normalizer, so the diff's
-  // verdict is exactly what the live filedrop connection would open (backslashes
-  // folded to forward slashes, all trailing slashes stripped, root-like paths
-  // preserved) -- no separate equality rule to drift from it. Either operand may
-  // be undefined: a split-directory config (inbound_path/outbound_path) carries
-  // no `path`, and a shared config carries no inbound/outbound, so undefined
-  // legitimately arrives. pushDirectoryConflicts calls this once per locator it
-  // compares (the single `path`, or each half of the split pair).
-  const norm = (p: string | undefined): string | undefined =>
-    p === undefined ? undefined : normalizeFiledropPath(p);
-  return norm(a) === norm(b);
-}
-
-/**
- * Append the directory-locator conflicts for a file-sync channel. A directory is
- * given either as a single shared path or as the split inbound/outbound pair;
- * this compares whichever form the `target` uses, so a split target is reconciled
- * pair-wise and a shared target by its single path (matching how the live
- * connection resolves each). An existing config in the other form differs in the
- * compared field (its value is unset), so a shared-vs-split mismatch is a
- * conflict like any other. `pathsEqual` is the channel's own path comparator and
- * `field` renders a config key to its snake_case message path. Only fields the
- * target actually sets are compared, so a locator the target leaves unset is not
- * a disagreement with whatever the config holds.
- */
-function pushDirectoryConflicts(
-  conflicts: ReconcileDiff[],
-  have: { path?: string; inboundPath?: string; outboundPath?: string },
-  want: { path?: string; inboundPath?: string; outboundPath?: string },
-  pathsEqual: (a: string | undefined, b: string | undefined) => boolean,
-  field: (key: "path" | "inbound_path" | "outbound_path") => string,
-): void {
-  // When the existing config is in the OTHER directory form than the target, the
-  // compared field is genuinely unset on the existing side -- but a bare
-  // "(unset)" hides the locator the config DOES hold in its own form, which an
-  // operator reads as "my config names no directory at all". Annotate the unset
-  // side with that locator so the conflict shows both forms. Only invoked when
-  // the field being rendered is actually unset, so it never fires for a
-  // same-form mismatch (where the existing value is shown directly).
-  const existingHint = (): string => {
-    if (have.path !== undefined)
-      return `${RECONCILE_UNSET} (the config uses a single shared path ${have.path})`;
-    if (have.inboundPath !== undefined || have.outboundPath !== undefined)
-      return (
-        `${RECONCILE_UNSET} (the config uses a split inbound_path ` +
-        `${have.inboundPath ?? RECONCILE_UNSET}, outbound_path ` +
-        `${have.outboundPath ?? RECONCILE_UNSET})`
-      );
-    return RECONCILE_UNSET;
-  };
-
-  const split =
-    want.inboundPath !== undefined || want.outboundPath !== undefined;
-  if (split) {
-    if (
-      want.inboundPath !== undefined &&
-      !pathsEqual(have.inboundPath, want.inboundPath)
-    )
-      conflicts.push({
-        field: field("inbound_path"),
-        existing: have.inboundPath ?? existingHint(),
-        incoming: want.inboundPath,
-      });
-    if (
-      want.outboundPath !== undefined &&
-      !pathsEqual(have.outboundPath, want.outboundPath)
-    )
-      conflicts.push({
-        field: field("outbound_path"),
-        existing: have.outboundPath ?? existingHint(),
-        incoming: want.outboundPath,
-      });
-    return;
-  }
-  if (want.path !== undefined && !pathsEqual(have.path, want.path))
-    conflicts.push({
-      field: field("path"),
-      existing: have.path ?? existingHint(),
-      incoming: want.path,
-    });
-}
-
-/**
- * Compare a pre-existing config's connection block against the connection the
- * online accept will actually use -- the {@link connectionFromURL} result, i.e.
- * the accept URL with any `--server-*` overrides already applied -- splitting
- * the disagreements into those that must abort the acceptance (`conflicts`) and
- * those that only warn (`warnings`).
- *
- * Comparing against that built `target` connection, rather than re-deriving the
- * effective values from the URL here, is deliberate: the diff's verdict then
- * matches what the live exchange does field for field. It cannot affirm a "match"
- * the connection later contradicts, and it inherits `connectionFromURL`'s own
- * encoding handling for free (so when that builder is taught to percent-decode
- * the path/userinfo, this comparison decodes with it).
- *
- * The split follows where vs how you reach the rendezvous. `host` and `path`
- * identify *which* drop you are meeting at; a mismatch there is almost always a
- * wrong-invitation or wrong-config paste, so it is a conflict and aborts before
- * any acceptance is sent. The channel (protocol), port, and credentials are
- * *how* you reach the same drop and are legitimately variable -- e.g. a file-sync
- * drop reached via `file://` by one party and `sftp://` by another, an alternate
- * SSH port or tunnel, or a different account -- so a mismatch warns and the run
- * proceeds: the live exchange uses the target, and the saved config is left
- * unchanged. Only fields the target actually sets are compared: a port, path, or
- * credential the target leaves unset (the URL omitted it and no override
- * supplied it) is not a disagreement with whatever the config holds. host is
- * compared case-insensitively and paths ignoring a trailing slash, matching how
- * DNS and FileSyncConnection treat them. Credential values are never echoed in a
- * warning -- a password or key in a log is a leak -- so those warnings report
- * only that the value differs. A channel mismatch short-circuits the per-channel
- * fields, which are not comparable across channels.
- *
- * @internal exported for testing
- */
-export function diffConnectionAgainstTarget(
-  existing: ConnectionConfig,
-  target: RunnableConnectionConfig,
-): { conflicts: ReconcileDiff[]; warnings: string[] } {
-  const conflicts: ReconcileDiff[] = [];
-  const warnings: string[] = [];
-
-  if (existing.channel !== target.channel) {
-    warnings.push(
-      `channel: specified ${target.channel}, saved ${existing.channel}`,
-    );
-    return { conflicts, warnings };
-  }
-
-  if (target.channel === "sftp") {
-    // Safe: existing.channel === target.channel === "sftp".
-    const have = (existing as SFTPConnectionConfig).server;
-    const want = target.server;
-
-    // host/path -> conflict (which drop you are meeting at). The directory is the
-    // single shared `server.path` or the split inbound/outbound pair, compared in
-    // whichever form the target uses.
-    if (!hostsEqual(have.host, want.host))
-      conflicts.push({
-        field: "connection.server.host",
-        existing: have.host,
-        incoming: want.host,
-      });
-    pushDirectoryConflicts(
-      conflicts,
-      have,
-      want,
-      sftpPathsEqual,
-      (key) => `connection.server.${key}`,
-    );
-
-    // port -> warn (how you reach the same host). An unset config port means the
-    // SFTP default, so a target restating that default is not a divergence.
-    if (
-      want.port !== undefined &&
-      want.port !== (have.port ?? DEFAULT_SFTP_PORT)
-    )
-      warnings.push(
-        `port: specified ${want.port}, saved ${have.port ?? "unset"}`,
-      );
-
-    // credentials -> warn, value never echoed.
-    if (want.username !== undefined && want.username !== have.username)
-      warnings.push("username: differs from the saved value");
-    if (want.password !== undefined && want.password !== have.password)
-      warnings.push("password: differs from the saved value");
-    if (want.privateKey !== undefined && want.privateKey !== have.privateKey)
-      warnings.push("private key: differs from the saved value");
-    if (
-      want.privateKeyPassphrase !== undefined &&
-      want.privateKeyPassphrase !== have.privateKeyPassphrase
-    )
-      warnings.push("private key passphrase: differs from the saved value");
-    // keyboard_interactive is a non-secret auth toggle (settable only via
-    // --server-keyboard-interactive, never from a URL), so its values are echoed
-    // like the port above rather than redacted like the credentials; the
-    // `!== undefined` guard keeps it silent unless the override actually set it.
-    if (
-      want.keyboardInteractive !== undefined &&
-      want.keyboardInteractive !== have.keyboardInteractive
-    )
-      warnings.push(
-        `keyboard-interactive: specified ${want.keyboardInteractive}, saved ` +
-          `${have.keyboardInteractive ?? "unset"}`,
-      );
-  } else if (target.channel === "filedrop") {
-    // filedrop's only locator is the directory -> conflict. No port/credentials
-    // apply. The directory is the single shared `path` or the split
-    // inbound/outbound pair, compared in whichever form the target uses.
-    pushDirectoryConflicts(
-      conflicts,
-      existing as FileDropConnectionConfig,
-      target,
-      filedropPathsEqual,
-      (key) => `connection.${key}`,
-    );
-  }
-  // webrtc never reaches here: connectionFromURL rejects a ws/wss URL before the
-  // target is built, so `target` is only ever sftp/filedrop, and a webrtc
-  // existing config is caught by the channel mismatch above.
-
-  return { conflicts, warnings };
 }
 
 /**
@@ -1449,705 +1109,5 @@ export function logOnlineBootstrapOutcome(
       `so you do not need to re-invite: recreate ${params.configFile} to match ` +
       `your connection and linkage settings before running a recurring ` +
       `'psilink exchange'. Keep the key file private.`,
-  );
-}
-
-// --- Command execution -------------------------------------------------------
-
-/**
- * Run a command body, mapping any thrown error to a process exit: a
- * {@link UsageError} to EX_USAGE (64), otherwise the error's own numeric
- * `exitCode` or EX_UNAVAILABLE (69). This is the single error->exit boundary for
- * the bootstrap commands; routing the whole handler body through it -- including
- * option parsing and the accept confirmation prompt -- means a thrown or
- * rejected step exits cleanly rather than crashing with an unhandled rejection.
- * The `?? exitCode` rung is load-bearing: `openInputSource` and `buildDataSpec`
- * throw plain `Error`s carrying `exitCode`, so a missing input file keeps its own
- * exit code rather than collapsing to 69.
- *
- * The error logger is created from `loggerName` lazily in the catch, so the body
- * is free to apply the configured log level (via `setDefaultLevel`) before
- * creating its own logger -- loglevel binds a logger's level at creation, so the
- * body's logger must be made after the level is set. `process.exit` is typed
- * `never`, so values produced inside `body` keep their definite-assignment
- * narrowing.
- */
-export async function runOrExit(
-  loggerName: string,
-  body: () => Promise<void>,
-): Promise<void> {
-  try {
-    await body();
-  } catch (err) {
-    getLogger(loggerName).error(sanitizeErrorForDisplay(err));
-    process.exit(
-      err instanceof UsageError
-        ? 64
-        : ((err as { exitCode?: number }).exitCode ?? 69),
-    );
-  }
-}
-
-// --- Shared CLI options ------------------------------------------------------
-
-/**
- * Per-command overrides for the descriptions of the common bootstrap options
- * whose accurate wording is command-specific -- chiefly whether the config/key
- * files are written or read, and whether the `server-*` (and `peer-id`)
- * overrides apply to a connection URL or to a loaded config. Keyed by the yargs
- * option name; an absent key keeps the default (`invite`/`accept`) wording. Only
- * the describe text varies -- the option name, type, alias, and default live
- * once in {@link addCommonBootstrapOptions}, so a new common flag is added there
- * alone and appears in every command.
- */
-export type CommonBootstrapDescribeOverrides = Partial<
-  Record<
-    | "config-file"
-    | "key-file"
-    | "server-port"
-    | "server-username"
-    | "server-password"
-    | "server-private-key"
-    | "server-private-key-passphrase"
-    | "server-keyboard-interactive"
-    | "peer-id"
-    | "timestamp-in-filename"
-    | "retain-files"
-    | "outbound-path",
-    string
-  >
->;
-
-/**
- * Add the options common to the bootstrap-style commands (config/key paths,
- * identity, SFTP credential overrides, connection/exchange tuning, logging,
- * recording). Positionals and command-specific options (`accept-timeout` for
- * `invite`, `save`/sweep controls for `zero-setup`/`exchange`) are added by each
- * command's own builder. `describe` lets a command supply accurate wording for
- * the few options whose meaning differs from the `invite`/`accept` default --
- * e.g. `exchange` reads a config and has no URL, so its `server-*` text names
- * the config rather than the URL.
- */
-export function addCommonBootstrapOptions(
-  cmd: Argv,
-  describe: CommonBootstrapDescribeOverrides = {},
-): Argv {
-  return cmd
-    .option("config-file", {
-      type: "string",
-      describe:
-        describe["config-file"] ??
-        `where to write psilink.yaml (default: ${DEFAULT_CONFIG_PATH})`,
-    })
-    .option("key-file", {
-      type: "string",
-      describe:
-        describe["key-file"] ??
-        `where to write .psilink.key (default: ${DEFAULT_KEY_PATH})`,
-    })
-    .option("identity", {
-      type: "string",
-      describe: "identity string for this party (name, org, contact)",
-    })
-    .option("server-port", {
-      type: "number",
-      describe:
-        describe["server-port"] ?? "server port; overrides the port in URL",
-    })
-    .option("server-username", {
-      type: "string",
-      describe:
-        describe["server-username"] ??
-        "server username; overrides the username in URL",
-    })
-    .option("server-password", {
-      type: "string",
-      describe:
-        describe["server-password"] ??
-        "server password; use @path to read from file; overrides the " +
-          "password in URL",
-    })
-    .option("server-private-key", {
-      type: "string",
-      describe:
-        describe["server-private-key"] ??
-        "SSH private key; use @path to read from file",
-    })
-    .option("server-private-key-passphrase", {
-      type: "string",
-      describe:
-        describe["server-private-key-passphrase"] ??
-        "passphrase for an encrypted SSH private key; use @path to read from " +
-          "file; requires --server-private-key",
-    })
-    .option("server-keyboard-interactive", {
-      type: "boolean",
-      describe:
-        describe["server-keyboard-interactive"] ??
-        "answer the server's keyboard-interactive prompts with the password; " +
-          "requires a password. Enable for a server that rejects the direct " +
-          "password method but accepts the same password over " +
-          "keyboard-interactive",
-    })
-    .option("connection-timeout", {
-      type: "string",
-      describe:
-        "how long to wait when connecting to the primary exchange server " +
-        `(maximum: ${MAX_TIMEOUT_SECONDS / 86_400}d). ` +
-        DURATION_VALUE_HELP,
-    })
-    .option("peer-timeout", {
-      alias: "t",
-      type: "string",
-      describe:
-        "how long to wait for the peer before giving up " +
-        `(maximum: ${MAX_TIMEOUT_SECONDS / 86_400}d). ` +
-        DURATION_VALUE_HELP,
-    })
-    .option("polling-frequency", {
-      type: "string",
-      describe:
-        "how often to poll the shared directory for the partner's files on " +
-        "the sftp/filedrop channels (default: 5s). A conservative default " +
-        "keeps within SFTP servers' anti-flood limits; a sub-second value is " +
-        "accepted for a demo against a controlled server but warns. " +
-        FINE_DURATION_VALUE_HELP,
-    })
-    .option("max-reconnect-attempts", {
-      type: "number",
-      describe: "maximum reconnection attempts before giving up; default: 3",
-    })
-    .option("log-level", {
-      type: "string",
-      describe: "silent | error | warn | info | debug | trace; default=info",
-    })
-    .option("log-file", {
-      type: "string",
-      describe:
-        "append all log output to this file instead of the terminal; the " +
-        "parent directory must already exist",
-    })
-    .option("record", {
-      type: "boolean",
-      default: true,
-      describe:
-        "after a successful exchange, write a self-attested audit record (a " +
-        "local artifact, not a signed receipt) and its private verification " +
-        "keys; use --no-record to skip",
-    })
-    .option("record-file", {
-      type: "string",
-      describe:
-        "path for the audit record (default: ./psilink-record-<timestamp>." +
-        "json); the private verification keys are written alongside it as " +
-        "<name>.keys.json",
-    })
-    .option("lockless-rendezvous", {
-      type: "boolean",
-      describe:
-        "use the ack-handshake rendezvous instead of the atomic lock-file " +
-        "race; required on sync-mediated transports that lack atomic " +
-        "exclusive-create or deletion visibility during rendezvous. Both " +
-        "parties must set this flag identically",
-    })
-    .option("peer-id", {
-      type: "string",
-      describe:
-        describe["peer-id"] ??
-        "stable identifier for this party; appears in filenames and logs. " +
-          "Requires timestamp_in_filename: true. Both parties must use " +
-          "distinct ids",
-    })
-    .option("timestamp-in-filename", {
-      type: "boolean",
-      describe:
-        describe["timestamp-in-filename"] ??
-        "encode a UTC timestamp and per-session counter in each outgoing " +
-          "message filename; --retain-files implies it. Both parties must use " +
-          "the same value",
-    })
-    .option("retain-files", {
-      type: "boolean",
-      describe:
-        describe["retain-files"] ??
-        "keep all exchange files as a permanent transcript instead of " +
-          "deleting them after consumption. Requires --timestamp-in-filename. " +
-          "Both parties must set this flag identically",
-    })
-    .option("outbound-path", {
-      type: "string",
-      describe:
-        describe["outbound-path"] ??
-        "use a separate outbound directory: the URL/positional path becomes " +
-          "the inbound (peer-written) directory and this is the outbound " +
-          "(self-written) directory, for managed shares and SFTP servers with " +
-          "distinct drop and pickup folders. Requires --retain-files; the two " +
-          "directories must differ. Leave unset for a single shared directory. " +
-          "Each party sets its own directories",
-    })
-    .option("verbose", {
-      alias: "v",
-      type: "count",
-      describe:
-        "generate additional logging information for sub-libraries at all " +
-        "logging levels",
-    });
-}
-
-/** The options common to `invite` and `accept`, parsed from yargs `Arguments`. */
-export interface CommonBootstrapOptions {
-  configFile: string;
-  keyFile: string;
-  identity?: string;
-  serverPort?: number;
-  serverUsername?: string;
-  serverPassword?: string;
-  serverPrivateKey?: string;
-  serverPrivateKeyPassphrase?: string;
-  serverKeyboardInteractive?: boolean;
-  connectionTimeout?: number;
-  peerTimeout?: number;
-  // The --polling-frequency override, in MILLISECONDS (not seconds like the two
-  // timeout fields above): the poll interval is millisecond-scaled and accepts a
-  // sub-second value, so it carries its native unit through to the connection's
-  // pollIntervalMs without a lossy seconds round-trip.
-  pollingFrequencyMs?: number;
-  maxReconnectAttempts?: number;
-  locklessRendezvous?: boolean;
-  peerId?: string;
-  timestampInFilename?: boolean;
-  retainFiles?: boolean;
-  outboundPath?: string;
-  record: boolean;
-  recordFile?: string;
-  logLevel: logLibrary.LogLevelNumbers;
-  logFile?: string;
-  verbosity: number;
-}
-
-/** Parse the {@link CommonBootstrapOptions} from yargs `Arguments`. */
-export function parseCommonBootstrapArgs(
-  argv: Arguments,
-): CommonBootstrapOptions {
-  const rawLogLevel = (
-    (singleValue(argv, "log-level") as string | undefined) || "info"
-  ).toLowerCase();
-  const logLevel = LOG_LEVELS[rawLogLevel];
-  if (logLevel === undefined)
-    throw new UsageError(`unrecognized log-level: ${argv["log-level"]}`);
-
-  // Each single-value (string/number) option is read through singleValue so a
-  // repeated flag is rejected with a clean usage error before its array value
-  // reaches a cast that lies about the type. The boolean and count options
-  // (lockless-rendezvous, timestamp-in-filename, retain-files, record, verbose)
-  // keep their plain casts: a repeat is valid for them.
-  return {
-    configFile:
-      (singleValue(argv, "config-file") as string | undefined) ??
-      DEFAULT_CONFIG_PATH,
-    keyFile:
-      (singleValue(argv, "key-file") as string | undefined) ?? DEFAULT_KEY_PATH,
-    identity: singleValue(argv, "identity") as string | undefined,
-    serverPort: singleValue(argv, "server-port") as number | undefined,
-    serverUsername: singleValue(argv, "server-username") as string | undefined,
-    // Credential values are carried through verbatim; an `@path` ref is read only
-    // at the live-use boundary (resolveConnectionCredentials in
-    // runOnlineBootstrap), so a persisted config keeps the `@path`, not the
-    // resolved secret.
-    serverPassword: singleValue(argv, "server-password") as string | undefined,
-    serverPrivateKey: singleValue(argv, "server-private-key") as
-      | string
-      | undefined,
-    serverPrivateKeyPassphrase: singleValue(
-      argv,
-      "server-private-key-passphrase",
-    ) as string | undefined,
-    // Boolean toggle, so it keeps a plain cast (a repeat is valid, like the other
-    // boolean flags); yargs yields true only when the enabling form is passed.
-    serverKeyboardInteractive: argv["server-keyboard-interactive"] as
-      | boolean
-      | undefined,
-    connectionTimeout: durationFlagSeconds(
-      argv,
-      "connection-timeout",
-      MAX_TIMEOUT_SECONDS,
-    ),
-    peerTimeout: durationFlagSeconds(argv, "peer-timeout", MAX_TIMEOUT_SECONDS),
-    // Read in milliseconds (durationFlagMs, not durationFlagSeconds) so a
-    // sub-second demo value survives; no product ceiling -- a large poll interval
-    // is only slow, and the schema field imposes no maximum (see durationFlagMs).
-    pollingFrequencyMs: durationFlagMs(argv, "polling-frequency"),
-    maxReconnectAttempts: nonNegativeIntFlag(
-      argv,
-      "max-reconnect-attempts",
-      MAX_RECONNECT_ATTEMPTS,
-    ),
-    locklessRendezvous: argv["lockless-rendezvous"] as boolean | undefined,
-    peerId: singleValue(argv, "peer-id") as string | undefined,
-    timestampInFilename: argv["timestamp-in-filename"] as boolean | undefined,
-    retainFiles: argv["retain-files"] as boolean | undefined,
-    outboundPath: singleValue(argv, "outbound-path") as string | undefined,
-    // yargs sets `record` to false on --no-record and true by the option's
-    // default otherwise, so it is always a boolean here.
-    record: argv["record"] as boolean,
-    recordFile: singleValue(argv, "record-file") as string | undefined,
-    logLevel,
-    logFile: singleValue(argv, "log-file") as string | undefined,
-    verbosity: (argv["verbose"] as number | undefined) ?? 0,
-  };
-}
-
-/**
- * The subset of parsed common options {@link connectionOverridesFrom} reads. A
- * `Pick` rather than the full {@link CommonBootstrapOptions} so the command
- * `Options` shapes that carry only the override fields (exchange's, zero-setup's)
- * also satisfy it; the full options object is assignable to it too.
- */
-export type ConnectionOverrideOptions = Pick<
-  CommonBootstrapOptions,
-  | "connectionTimeout"
-  | "peerTimeout"
-  | "pollingFrequencyMs"
-  | "maxReconnectAttempts"
-  | "serverUsername"
-  | "serverPassword"
-  | "serverPrivateKey"
-  | "serverPrivateKeyPassphrase"
-  | "serverKeyboardInteractive"
-  | "serverPort"
-  | "locklessRendezvous"
-  | "peerId"
-  | "timestampInFilename"
-  | "retainFiles"
-  | "outboundPath"
->;
-
-/**
- * Map the parsed (flat) common options to the structured connection-override
- * shape, fanning the `server-*`/`--outbound-path` flags into the server
- * sub-group and the timeout/toggle flags into the options sub-group. The
- * synthesized `extra.peerTimeout` (from `--accept-timeout` on the online invite
- * path) belongs to the options sub-group and takes precedence over the parsed
- * `--peer-timeout`.
- */
-export function connectionOverridesFrom(
-  options: ConnectionOverrideOptions,
-  extra: { peerTimeout?: number } = {},
-): ConnectionOverrides {
-  return {
-    server: {
-      username: options.serverUsername,
-      password: options.serverPassword,
-      privateKey: options.serverPrivateKey,
-      privateKeyPassphrase: options.serverPrivateKeyPassphrase,
-      keyboardInteractive: options.serverKeyboardInteractive,
-      port: options.serverPort,
-      outboundPath: options.outboundPath,
-    },
-    options: {
-      connectionTimeout: options.connectionTimeout,
-      peerTimeout: extra.peerTimeout ?? options.peerTimeout,
-      // Already in milliseconds -- fed straight into the connection's
-      // pollIntervalMs by applyConnectionOverrides, with no seconds scaling.
-      pollIntervalMs: options.pollingFrequencyMs,
-      maxReconnectAttempts: options.maxReconnectAttempts,
-      locklessRendezvous: options.locklessRendezvous,
-      peerId: options.peerId,
-      timestampInFilename: options.timestampInFilename,
-      retainFiles: options.retainFiles,
-    },
-  };
-}
-
-/**
- * Warn that the file-sync-only flags (`--lockless-rendezvous`, `--retain-files`,
- * `--polling-frequency`) have no effect on a channel that is not `sftp` or
- * `filedrop`, naming whichever flags the caller actually set. The channel is
- * taken as input so the one helper serves both call sites: `exchange` derives it
- * from the loaded connection (post-override), `zero-setup` from the server URL
- * (pre-connection). A file-sync channel warns for none of them. Shared so the
- * wording cannot drift between the two commands.
- *
- * `--polling-frequency` belongs here because `pollIntervalMs` is a FileSyncOptions
- * field {@link applyConnectionOverrides} applies only on `sftp`/`filedrop`, so on
- * a non-file-sync channel the override is dropped exactly as the two toggles are;
- * this is where its ignored-flag warning lives, and {@link warnLowPollingFrequency}
- * (the aggressively-low advisory) is correspondingly a no-op off those channels so
- * the two never both fire.
- *
- * `--outbound-path` is deliberately NOT one of these flags: unlike the silently-
- * ignored options above, it is a hard error on a non-file-sync channel (the
- * URL-driven commands reject a webrtc URL before overrides apply, and
- * applyConnectionOverrides throws on a webrtc config), so it needs no
- * "ignored" warning -- a warning here would falsely promise it was tolerated.
- */
-export function warnUnsupportedFileSyncFlags(
-  channel: ConnectionConfig["channel"],
-  flags: {
-    locklessRendezvous?: boolean;
-    retainFiles?: boolean;
-    pollingFrequencyMs?: number;
-  },
-  log: { warn: (message: string) => void },
-): void {
-  if (channel === "sftp" || channel === "filedrop") return;
-  if (flags.locklessRendezvous === true)
-    log.warn(
-      `--lockless-rendezvous has no effect on the ${channel} channel and will ` +
-        "be ignored; it is only supported on sftp and filedrop",
-    );
-  if (flags.retainFiles === true)
-    log.warn(
-      `--retain-files has no effect on the ${channel} channel and will be ` +
-        "ignored; it is only supported on sftp and filedrop",
-    );
-  // A number gate (not `=== true`): pollingFrequencyMs is set or unset, with no
-  // negated CLI form to fold to a default the way the boolean toggles have.
-  if (flags.pollingFrequencyMs !== undefined)
-    log.warn(
-      `--polling-frequency has no effect on the ${channel} channel and will be ` +
-        "ignored; it is only supported on sftp and filedrop",
-    );
-}
-
-/**
- * Millisecond threshold below which a `--polling-frequency` override draws the
- * anti-flood warning ({@link warnLowPollingFrequency}). One second: at or above
- * it a poll cadence stays within the anti-flood budgets commercial SFTP servers
- * enforce, so it is silent; below it the operator is warned but not blocked (a
- * demo against a controlled server legitimately wants a sub-second poll).
- */
-export const LOW_POLLING_FREQUENCY_WARN_MS = 1000;
-
-/**
- * Warn -- but do not block -- when a `--polling-frequency` override is set
- * aggressively low (below {@link LOW_POLLING_FREQUENCY_WARN_MS}), because a
- * sub-second poll hammers the shared directory with listings and can trip a
- * commercial SFTP server's anti-flood/DoS protection and drop the connection (the
- * partner-deployment failure that motivated the conservative default). There is
- * deliberately no hard floor: a demo against a controlled server may want ~100ms.
- * A no-op when the flag is absent or at/above the threshold, so a conservative
- * value emits nothing.
- *
- * A no-op on a non-file-sync `channel`: `pollIntervalMs` is applied only on
- * `sftp`/`filedrop` (see {@link applyConnectionOverrides}), so on any other
- * channel the override is dropped and its ignored-flag warning is emitted by
- * {@link warnUnsupportedFileSyncFlags} instead -- the aggressively-low advisory
- * here would be misleading, since no poll ever runs at that rate. `channel` is
- * `undefined` when the command could not resolve one (an unknown URL scheme in
- * zero-setup), which is likewise not a file-sync channel, so it no-ops too.
- *
- * Scoped to the CLI override value (the parsed `pollingFrequencyMs`), not the
- * effective merged interval: a low value already sitting in a loaded config's
- * `pollIntervalMs` is the operator's committed choice and is not re-litigated on
- * every run. The interpolated value is the operator's own numeric flag argument
- * (non-secret), so echoing it is safe. Called on the paths where the override
- * actually takes effect (zero-setup, `exchange`, and the ONLINE invite/accept);
- * the offline paths route `--polling-frequency` through
- * {@link warnOptionsOverridesIgnoredOffline} instead, since it is dropped there.
- */
-export function warnLowPollingFrequency(
-  channel: ConnectionConfig["channel"] | undefined,
-  pollingFrequencyMs: number | undefined,
-  log: { warn: (message: string) => void },
-): void {
-  if (channel !== "sftp" && channel !== "filedrop") return;
-  if (
-    pollingFrequencyMs === undefined ||
-    pollingFrequencyMs >= LOW_POLLING_FREQUENCY_WARN_MS
-  )
-    return;
-  log.warn(
-    `--polling-frequency ${pollingFrequencyMs}ms is below ` +
-      `${LOW_POLLING_FREQUENCY_WARN_MS}ms; polling this aggressively may trip ` +
-      "an SFTP server's anti-flood/DoS protection and drop the connection. Use " +
-      "a sub-second interval only against a controlled server (for example a demo).",
-  );
-}
-
-/**
- * The server-block overrides that have no effect on an OFFLINE invite/accept,
- * read by {@link warnServerOverridesIgnoredOffline}. A structural subset of
- * {@link CommonBootstrapOptions} (which is assignable to it) holding the
- * `--server-*` flags and `--outbound-path`: the overrides {@link
- * applyConnectionOverrides} writes into `connection.server` (host/port/
- * credentials) and the channel's directory paths -- the connection's address and
- * credentials -- as opposed to the tuning/behavior fields it writes into
- * `connection.options`. This server block is what the offline placeholder stands
- * in for: {@link connectionFromEndpoint} writes a `REPLACE_WITH_...` host/username
- * stub, or a seeded host/port/path, so these are the overrides that would have
- * populated a field the operator edits. (The set carries the
- * username/password/private-key the placeholder marks for replacement, so it is
- * NOT the credential-free public {@link ConnectionEndpoint}, which omits them by
- * construction.)
- *
- * The tuning overrides (timeouts, `--max-reconnect-attempts`) and the file-sync
- * toggles (`--retain-files`, `--peer-id`, etc.) are ALSO dropped offline, but they
- * target the `connection.options` tuning/behavior fields, which the offline flow
- * never populates from an override -- a split seed pre-seeds only the fixed
- * retain-mode trio there ({@link SPLIT_SEED_OPTIONS}), never operator tuning. So
- * the right diagnostic for them is a differently-worded one ("set it under
- * connection.options"), tracked as a follow-up, not folded in here where it would
- * blur this message's "set the connection details in that block" remedy.
- */
-export type OfflineIgnoredServerOverrides = Pick<
-  CommonBootstrapOptions,
-  | "serverUsername"
-  | "serverPassword"
-  | "serverPrivateKey"
-  | "serverPrivateKeyPassphrase"
-  | "serverKeyboardInteractive"
-  | "serverPort"
-  | "outboundPath"
->;
-
-/**
- * Warn that the server-block overrides (`--server-username`, `--server-password`,
- * `--server-private-key`, `--server-private-key-passphrase`,
- * `--server-keyboard-interactive`, `--server-port`, and `--outbound-path`) have no
- * effect
- * on an OFFLINE invite/accept. Those paths write a placeholder (invite) or
- * invitation-endpoint-seeded (accept) connection block for the operator to edit
- * before `psilink exchange`, rather than building a connection from a URL the way
- * the online and zero-setup paths do -- so they go through {@link
- * connectionFromEndpoint}, which applies no connection overrides, and these flags
- * would otherwise be parsed and silently dropped. The warning names exactly the
- * flags the operator set; it is a no-op when none are set. Scoped to the server
- * block these overrides populate (see {@link OfflineIgnoredServerOverrides});
- * shared between invite and accept so the wording cannot drift, and so the whole
- * `--server-*`/`--outbound-path` set is treated uniformly rather than only the one
- * flag (`--outbound-path`) whose silent loss is the most surprising.
- */
-export function warnServerOverridesIgnoredOffline(
-  options: OfflineIgnoredServerOverrides,
-  log: { warn: (message: string) => void },
-): void {
-  // Security invariant: push only the flag NAME, never the override value. A
-  // --server-password / --server-private-key / --server-private-key-passphrase
-  // value may be an inline secret (or an unresolved @path), and this warning
-  // reaches the terminal and any --log-file; interpolating a value here would
-  // leak it. Keep the emitted message static apart from this flag-name list.
-  const ignored: string[] = [];
-  if (options.serverUsername !== undefined) ignored.push("--server-username");
-  if (options.serverPassword !== undefined) ignored.push("--server-password");
-  if (options.serverPrivateKey !== undefined)
-    ignored.push("--server-private-key");
-  if (options.serverPrivateKeyPassphrase !== undefined)
-    ignored.push("--server-private-key-passphrase");
-  // Gate on `=== true`, not presence: yargs sets the negated form
-  // (--no-server-keyboard-interactive) to false, which is also the default, so
-  // only the enabling form was an override that could have done anything.
-  if (options.serverKeyboardInteractive === true)
-    ignored.push("--server-keyboard-interactive");
-  if (options.serverPort !== undefined) ignored.push("--server-port");
-  if (options.outboundPath !== undefined) ignored.push("--outbound-path");
-  if (ignored.length === 0) return;
-  log.warn(
-    `${ignored.join(", ")} ${ignored.length === 1 ? "has" : "have"} no effect ` +
-      "on an offline invite/accept: the connection block is written for you to " +
-      "edit (a placeholder, or seeded from the invitation endpoint), not built " +
-      "from a URL. Set the connection details directly in that block -- the " +
-      "server host/port/credentials, or the inbound_path/outbound_path split " +
-      "directory -- before running 'psilink exchange', or pass these flags on " +
-      "an online invite/accept, the zero-setup exchange, or 'psilink exchange'.",
-  );
-}
-
-/**
- * The connection-OPTIONS overrides that have no effect on an OFFLINE
- * invite/accept, read by {@link warnOptionsOverridesIgnoredOffline}. A structural
- * subset of {@link CommonBootstrapOptions} (assignable to it) holding the
- * tuning/toggle flags {@link applyConnectionOverrides} writes into
- * `connection.options`: the SharedOptions timeouts/reconnect bound
- * (`--connection-timeout`, `--peer-timeout`, `--max-reconnect-attempts`), the
- * file-sync poll interval (`--polling-frequency`), and the file-sync toggles
- * (`--lockless-rendezvous`, `--peer-id`, `--retain-files`,
- * `--timestamp-in-filename`) -- HOW the exchange behaves, as opposed to the
- * server block's WHERE/credentials that {@link OfflineIgnoredServerOverrides}
- * covers.
- *
- * The offline placeholder has no `options` block on any channel (a split seed
- * pre-seeds only the fixed retain-mode trio there, {@link SPLIT_SEED_OPTIONS},
- * never operator tuning), so these overrides are parsed and silently dropped on
- * the offline paths. They are the `connection.options` half deliberately scoped
- * out of {@link OfflineIgnoredServerOverrides}: warning for them belongs
- * in a separate, differently-worded diagnostic ("set them under
- * connection.options") rather than folded into the server warning's "set the
- * connection details in that block" remedy.
- *
- * Note the synthesized `--accept-timeout` is NOT here: it feeds the override
- * bag's `peerTimeout` only through {@link connectionOverridesFrom}'s `extra` arg
- * on the ONLINE invite path. This read is of the parsed `peerTimeout` field
- * (`--peer-timeout`) directly, so an offline invite with `--accept-timeout` but
- * no `--peer-timeout` does not warn spuriously.
- */
-export type OfflineIgnoredOptionsOverrides = Pick<
-  CommonBootstrapOptions,
-  | "connectionTimeout"
-  | "peerTimeout"
-  | "pollingFrequencyMs"
-  | "maxReconnectAttempts"
-  | "locklessRendezvous"
-  | "peerId"
-  | "timestampInFilename"
-  | "retainFiles"
->;
-
-/**
- * Warn that the connection-OPTIONS overrides (`--connection-timeout`,
- * `--peer-timeout`, `--polling-frequency`, `--max-reconnect-attempts`,
- * `--lockless-rendezvous`, `--peer-id`, `--timestamp-in-filename`,
- * `--retain-files`) have no effect on an
- * OFFLINE invite/accept. Like the server-block overrides those paths go through
- * {@link connectionFromEndpoint}, which applies no connection overrides, but
- * these target `connection.options` -- a block the offline placeholder does not
- * contain on any channel -- so they would otherwise be parsed and silently
- * dropped. The warning names exactly the flags the operator set; it is a no-op
- * when none are set.
- *
- * Kept distinct from {@link warnServerOverridesIgnoredOffline}: that one's remedy
- * is "set the connection details in that block" (host/port/credentials/paths),
- * whereas these land under `connection.options`, so the remedy wording points
- * there instead. Shared between invite and accept so the wording cannot drift.
- */
-export function warnOptionsOverridesIgnoredOffline(
-  options: OfflineIgnoredOptionsOverrides,
-  log: { warn: (message: string) => void },
-): void {
-  // Security invariant: push only the flag NAME, never the override value. This
-  // warning reaches the terminal and any --log-file, so interpolating a value
-  // here would leak it. None of these flags is a secret today, but --peer-id is
-  // an operator-supplied free string and the set may grow, so keep the emitted
-  // message static apart from this flag-name list -- the same discipline
-  // warnServerOverridesIgnoredOffline holds for its --server-password/-key.
-  const ignored: string[] = [];
-  if (options.connectionTimeout !== undefined)
-    ignored.push("--connection-timeout");
-  if (options.peerTimeout !== undefined) ignored.push("--peer-timeout");
-  if (options.pollingFrequencyMs !== undefined)
-    ignored.push("--polling-frequency");
-  if (options.maxReconnectAttempts !== undefined)
-    ignored.push("--max-reconnect-attempts");
-  // The three toggles gate on `=== true`, not presence: yargs sets the negated
-  // form (--no-retain-files etc.) to `false`, and `false` is the default a fresh
-  // placeholder already carries, so only the enabling form was an override that
-  // could have done something. This matches warnUnsupportedFileSyncFlags's
-  // `=== true` gate on the same toggles and avoids a warning that names
-  // --retain-files when the operator actually typed --no-retain-files.
-  if (options.locklessRendezvous === true)
-    ignored.push("--lockless-rendezvous");
-  if (options.peerId !== undefined) ignored.push("--peer-id");
-  if (options.timestampInFilename === true)
-    ignored.push("--timestamp-in-filename");
-  if (options.retainFiles === true) ignored.push("--retain-files");
-  if (ignored.length === 0) return;
-  log.warn(
-    `${ignored.join(", ")} ${ignored.length === 1 ? "has" : "have"} no effect ` +
-      "on an offline invite/accept: the connection block is written for you to " +
-      "edit (a placeholder, or seeded from the invitation endpoint), and these " +
-      "tuning options are not applied to it. Set them under connection.options " +
-      "in the written config before running 'psilink exchange', or pass these " +
-      "flags on an online invite/accept, the zero-setup exchange, or 'psilink " +
-      "exchange'.",
   );
 }
