+++ apps/cli/test/unit/logFile.test.ts
@@ -6,7 +6,7 @@ import logLibrary from "loglevel";
 import { getDiagnosticSink, getLogger, UsageError } from "@psilink/core";
 
 import { configureLogFile } from "../../src/util/cli";
-import { parseCommonBootstrapArgs } from "../../src/commands/bootstrap";
+import { parseCommonBootstrapArgs } from "../../src/optionDefinitions";
 import {
   argv,
   captureStdio,
